requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
exceptiongroup==1.1.3
fastapi==0.103.1
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.17.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.5
outcome==1.2.0
packaging==23.1
proto-plus==1.22.3
protobuf==4.24.3
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
starlette==0.27.0
termcolor==2.3.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.23.2
webdriver-manager==4.0.0
wsproto==1.2.0
